feat: add initial project structure and core functionality for Takharrujy platform
- Introduce essential files including .dockerignore and .gitignore for build and collaboration management
- Implement pom.xml for Maven project configuration
- Create initial Java classes for application structure, including main application entry point and service classes
- Establish domain entities for user, project, and department management
- Set up repository interfaces for data access and manipulation
- Add configuration classes for database and security settings
- Implement email service and token management for user authentication
- Include validation mechanisms for Arabic text input

This commit lays the foundation for the Takharrujy platform, enabling further development of user profile management, project tracking, and collaboration features.

+++ src/main/java/com/university/takharrujy/application/service/VirusScanService.java
@@ -25,7 +25,7 @@ public class VirusScanService {
         // Example virus signatures - in production use real antivirus engine
         new byte[]{0x4D, 0x5A}, // PE executable header
         new byte[]{0x7F, 0x45, 0x4C, 0x46}, // ELF header
-        new byte[]{0xCA, 0xFE, 0xBA, 0xBE} // Java class file
+        new byte[]{(byte) 0xCA, (byte) 0xFE, (byte) 0xBA, (byte) 0xBE} // Java class file
     );
     
     // Dangerous file extensions
